+++ CookManagement.VSA/Features/Users/GetUserRecords/UserRecordsResponse.cs
@@ -5,13 +5,13 @@
         public int UserId { get; init; }
         public string ProductCode { get; init; } = String.Empty;
         public string ProductName { get; init; } = String.Empty;
-        public int InitialInventory { get; init; } = 0;
-        public int FinalInventory { get; init; } = 0;
-        public int Difference { get; init; } = 0;
-        public int DailyMove { get; init; } = 0;
+        public Double InitialInventory { get; init; } = 0;
+        public Double FinalInventory { get; init; } = 0;
+        public Double Difference { get; init; } = 0;
+        public Double DailyMove { get; init; } = 0;
         public int Entries { get; init; } = 0;
-        public int Courtesy { get; init; } = 0;
-        public int Damaged { get; init; } = 0;
-        public int Remains { get; init; } = 0;
+        public Double Courtesy { get; init; } = 0;
+        public Double Damaged { get; init; } = 0;
+        public Double Remains { get; init; } = 0;
     }
 }
